# perf: vttablet/mysql optimizations