--- a/go/bucketpool/bucketpool.go
+++ b/go/bucketpool/bucketpool.go
@@ -17,7 +17,7 @@ limitations under the License.
 package bucketpool
 
 import (
-	"math"
+	"math/bits"
 	"sync"
 )
 
@@ -69,12 +69,10 @@ func (p *Pool) findPool(size int) *sizedPool {
 	if size > p.maxSize {
 		return nil
 	}
-	idx := int(math.Ceil(math.Log2(float64(size) / float64(p.minSize))))
-	if idx < 0 {
-		idx = 0
-	}
-	if idx > len(p.pools)-1 {
-		return nil
+	div, rem := bits.Div64(0, uint64(size), uint64(p.minSize))
+	idx := bits.Len64(div)
+	if rem == 0 && div != 0 && (div&(div-1)) == 0 {
+		idx = idx - 1
 	}
 	return p.pools[idx]
 }